move state matching to it's own file
